+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.20.0...HEAD) - TBD
 
+### :racing_car: Performance
+
+- Skip coverage boundary-length variants that are structurally impossible for the declared `pattern`.
+
 ## [4.20.0](https://github.com/schemathesis/schemathesis/compare/v4.19.0...v4.20.0) - 2026-05-24
 
 ### :rocket: Added